moved register routes to register.py
