+++ routes/__init__.py
@@ -1,1 +1,2 @@
 from .home import home_bp
+from .register import register_bp
