+++ templates/login.html
@@ -26,7 +26,7 @@
     </a>
 </div>
 
-<p>Don't have an account? <a href="{{ url_for('register') }}">Register here</a></p>
+<p>Don't have an account? <a href="{{ url_for('register.register') }}">Register here</a></p>
 <p>If you're admin, <a href="{{ url_for('signingoogle') }}">Sign in here.</a></p>
 
 <script>
