+++ templates/register.html
@@ -2,7 +2,7 @@
 
 {% block content %}
 <h2>Register</h2>
-<form method="POST" action="{{ url_for('register') }}">
+<form method="POST" action="{{ url_for('register.register') }}">
 
     <label for="firstname">First Name</label>
     <input type="text" id="firstname" name="firstname" placeholder="First Name" required>
